Filter to control running the hook sending SMS
Options to change $run_hook variable (bool). Default true to proceed with, but can be override via filter "run_woosms_hook_changeOrderStatusHook" not to run, if necessary.

Reason: Some orders are specifically created with purpose not to send info, but plugin does't allow filtration. This filter allows to make that option.